fixing tests that broke because of refactoring
