+++ src/sdk/python/rtdip_sdk/pipelines/data_quality/data_manipulation/spark/__init__.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 from .normalization import *
-from .prediction import *
 from .dimensionality_reduction import DimensionalityReduction
 from .duplicate_detection import DuplicateDetection
 from .interval_filtering import IntervalFiltering
