+++ src/sdk/python/rtdip_sdk/pipelines/forecasting/spark/arima.py
@@ -29,9 +29,9 @@ from regex import regex
 from statsmodels.tsa.arima.model import ARIMA
 import numpy as np
 
-from ...interfaces import DataManipulationBaseInterface
+from ...data_quality.data_manipulation.interfaces import DataManipulationBaseInterface
 from ...data_quality.input_validator import InputValidator
-from ...._sdk_utils import _prepare_pandas_to_convert_to_spark
+from ...._sdk_utils.pandas import _prepare_pandas_to_convert_to_spark
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.models import (
     Libraries,
     SystemType,
@@ -61,7 +61,7 @@ class ArimaPrediction(DataManipulationBaseInterface, InputValidator):
     import pandas
     from pyspark.sql import SparkSession
 
-    from rtdip_sdk.pipelines.data_quality.data_manipulation.spark.prediction.arima import ArimaPrediction
+    from rtdip_sdk.pipelines.data_quality.forecasting.spark.arima import ArimaPrediction
 
     import rtdip_sdk.pipelines._pipeline_utils.spark as spark_utils
