+++ tests/sdk/python/rtdip_sdk/pipelines/forecasting/spark/test_linear_regression.py
@@ -23,7 +23,7 @@ from pyspark.sql.types import (
     FloatType,
 )
 from datetime import datetime
-from src.sdk.python.rtdip_sdk.pipelines.machine_learning.spark.linear_regression import (
+from src.sdk.python.rtdip_sdk.pipelines.forecasting.spark.linear_regression import (
     LinearRegression,
 )
 from src.sdk.python.rtdip_sdk.pipelines.transformers.spark.machine_learning.columns_to_vector import (
